CLJS-1226: Add on-testing-complete hook
The function cljs.test/*on-testing-complete-fn* is called after all tests are done and it takes the test summary as its only argument.
It can be set by the user by calling cljs.test/set-on-testing-complete! which is exported to JS (via ^:export).

Additional: ^:exported was also added to cljs.test/successful? to be called from JS test runners.